feat(db) implement config secrets referencing
### Summary

This commit adds auto-dereferencing of `string` fields that are `referenceable=true`.
References are resolved on proxy nodes when reading entities.

The references cannot be validated on control planes, thus some validations need
to be disabled:

- Field validation rules cannot be applied if field value is a reference
- Entity validation rules cannot be applied if any of the fields involved contains
  references

This is only enabled on string fields and arrays and sets that have elements of string
type.